src/procap/cloud/requirements.txt
azure-ai-ml
azure-identity
fastapi
uvicorn
pydantic
torch
transformers